refactor(memory): Move the memory statistics code into its own file, separate allocation-specific tags from arena-specific tags.
- Implement thread memory attachments for windows (using TLS)
- Refactor tests to take the changes into account

+++ src/platform/windows/mc_thread_windows.h
@@ -9,6 +9,6 @@ typedef struct {
     struct ThreadInternal* internal;
 } MCThread;
 
-typedef i32 MCThreadKey;
+typedef i64 MCThreadKey;
 
 #endif /* ! MCTHREAD_WINDOWS_H */
